requirements.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
accelerate==0.18.0
aiohttp @ file:///tmp/build/80754af9/aiohttp_1646806366512/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
antlr4-python3-runtime==4.9.3
async-timeout @ file:///tmp/build/80754af9/async-timeout_1637851218186/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
beautifulsoup4==4.12.2
blinker==1.4
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1646038465422/work
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
cmake==3.26.1
contourpy==1.0.6
cryptography @ file:///tmp/build/80754af9/cryptography_1652083738073/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
docker-pycreds==0.4.0
einops==0.5.0
filelock==3.9.0
fonttools==4.25.0
frozenlist @ file:///tmp/build/80754af9/frozenlist_1637767111923/work
fsspec==2023.4.0
ftfy==6.1.1
future==0.18.2
gitdb==4.0.10
GitPython==3.1.30
google-auth @ file:///opt/conda/conda-bld/google-auth_1646735974934/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
googledrivedownloader @ file:///home/conda/feedstock_root/build_artifacts/googledrivedownloader_1619807768586/work
grpcio @ file:///tmp/build/80754af9/grpcio_1637590823556/work
huggingface-hub==0.12.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562408398/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1653292039266/work
kornia==0.6.11
lightning-utilities==0.8.0
lit==16.0.0
Markdown @ file:///tmp/build/80754af9/markdown_1614363528767/work
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.6.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpmath==1.3.0
multidict @ file:///opt/conda/conda-bld/multidict_1662369340274/work
munkres==1.1.4
networkx @ file:///opt/conda/conda-bld/networkx_1657784097507/work
numexpr @ file:///opt/conda/conda-bld/numexpr_1656940300424/work
numpy @ file:///tmp/abs_653_j00fmm/croots/recipe/numpy_and_numpy_base_1659432701727/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib @ file:///tmp/abs_08ngfezid4/croots/recipe/oauthlib_1659642459222/work
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.3
pathtools==0.1.2
Pillow==9.2.0
ply==3.11
promise==2.3
protobuf==3.20.1
psutil==5.9.2
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///opt/conda/conda-bld/pyjwt_1657544592787/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5-sip==12.11.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-louvain @ file:///tmp/build/80754af9/python-louvain_1612304551119/work
pytz @ file:///opt/conda/conda-bld/pytz_1654762638606/work
PyYAML==6.0
regex==2022.10.31
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-learn @ file:///tmp/abs_d76175bc-917a-47d4-9994-b56265948a6328vmoe2o/croots/recipe/scikit-learn_1658419412415/work
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1653073867187/work
sentencepiece==0.1.99
sentry-sdk==1.12.1
setproctitle==1.3.2
shortuuid==1.0.11
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
soupsieve==2.4.1
sympy==1.11.1
tensorboard @ file:///home/builder/stiwari/miniconda3/envs/tf_new_env/conda-bld/tensorboard_1661447826088/work/tensorboard-2.9.0-py3-none-any.whl
tensorboard-data-server @ file:///tmp/build/80754af9/tensorboard-data-server_1633035064162/work/tensorboard_data_server-0.6.0-py3-none-manylinux2010_x86_64.whl
tensorboard-plugin-wit @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard-plugin-wit_1658918494740/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tokenizers==0.13.2
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==1.13.1+cu116
torch-cluster @ file:///usr/share/miniconda/envs/test/conda-bld/pytorch-cluster_1631029005429/work
torch-geometric @ file:///usr/share/miniconda/envs/test/conda-bld/pyg_1640156451028/work
torch-scatter @ file:///usr/share/miniconda/envs/test/conda-bld/pytorch-scatter_1634900577572/work
torch-sparse @ file:///usr/share/miniconda/envs/test/conda-bld/pytorch-sparse_1631173533284/work
torch-spline-conv @ file:///usr/share/miniconda/envs/test/conda-bld/pytorch-spline-conv_1631007898768/work
torchaudio==0.13.1+cu116
torchmetrics==0.11.4
torchvision==0.14.1+cu116
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1650891076910/work
transformers==4.26.0
triton==2.0.0
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3 @ file:///tmp/abs_5dhwnz6atv/croots/recipe/urllib3_1659110457909/work
wandb==0.13.7
wcwidth==0.2.5
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
yacs @ file:///tmp/build/80754af9/yacs_1634047592950/work
yarl @ file:///opt/conda/conda-bld/yarl_1661437085904/work
zipp @ file:///opt/conda/conda-bld/zipp_1652341764480/work
